# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Finance App API",
    "description": "API para gestão de fluxo de caixa, contas bancárias e investimentos.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/categories": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "List Categories",
        "operationId": "list_categories_api_categories_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CategoryResponse"
                  },
                  "type": "array",
                  "title": "Response List Categories Api Categories Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Categories"
        ],
        "summary": "Create Category",
        "operationId": "create_category_api_categories_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/categories/{category_id}": {
      "patch": {
        "tags": [
          "Categories"
        ],
        "summary": "Update Category",
        "description": "Edição parcial. Renomear categoria em uso é permitido (decisão C11).",
        "operationId": "update_category_api_categories__category_id__patch",
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Category Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Categories"
        ],
        "summary": "Delete Category",
        "description": "Exclui a categoria; **409** se ela estiver em uso (decisões C9/C10).\n\nAs duas FKs para `categories.id` são `ondelete=\"RESTRICT\"`, então o banco já\nrecusaria — mas via `IntegrityError`, que sobe como 500. A checagem\nantecipada é o que transforma isso num conflito legível.\n\nOs dois vínculos são verificados: uma categoria pode estar presa a um\nparcelamento **sem ter transação nenhuma**, e olhar só `Transaction`\ndeixaria esse caso escapar para o 500.",
        "operationId": "delete_category_api_categories__category_id__delete",
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Category Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/installments": {
      "get": {
        "tags": [
          "Installments"
        ],
        "summary": "List Installments",
        "operationId": "list_installments_api_installments_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InstallmentResponse"
                  },
                  "type": "array",
                  "title": "Response List Installments Api Installments Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Installments"
        ],
        "summary": "Create Installment",
        "operationId": "create_installment_api_installments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstallmentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstallmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/installments/{installment_id}": {
      "patch": {
        "tags": [
          "Installments"
        ],
        "summary": "Update Installment",
        "description": "Edição parcial. Não toca saldo — parcelamento nunca tocou.\n\n`current_installment` pode ultrapassar `total_installments`: é o estado\n\"quitado\" (decisão D13), não um erro de validação.",
        "operationId": "update_installment_api_installments__installment_id__patch",
        "parameters": [
          {
            "name": "installment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Installment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstallmentUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstallmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CategoryCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 50,
            "title": "Name",
            "description": "Nome da categoria"
          },
          "icon_name": {
            "type": "string",
            "maxLength": 50,
            "title": "Icon Name",
            "description": "Nome do ícone da Lucide"
          },
          "budget": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Budget",
            "description": "Orçamento mensal para a categoria",
            "default": "0.00"
          },
          "color": {
            "type": "string",
            "maxLength": 50,
            "title": "Color",
            "description": "Cor em formato OKLCH ou HEX"
          }
        },
        "type": "object",
        "required": [
          "name",
          "icon_name",
          "color"
        ],
        "title": "CategoryCreate"
      },
      "CategoryRef": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "color": {
            "type": "string",
            "title": "Color"
          },
          "icon_name": {
            "type": "string",
            "title": "Icon Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "color",
          "icon_name"
        ],
        "title": "CategoryRef",
        "description": "Categoria aninhada nas responses de transação e parcelamento.\n\nCarrega `color` e `icon_name` junto do nome para que o front pinte a badge\nsem uma segunda chamada a /categories. Serialização direta do ORM pela\nrelação — ver \"Design Patterns\" no CLAUDE.md; toda listagem que expõe este\ncampo precisa de `joinedload`, senão vira N+1."
      },
...
      "CategoryResponse": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 50,
            "title": "Name",
            "description": "Nome da categoria"
          },
          "icon_name": {
            "type": "string",
            "maxLength": 50,
            "title": "Icon Name",
            "description": "Nome do ícone da Lucide"
          },
          "budget": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Budget",
            "description": "Orçamento mensal para a categoria",
            "default": "0.00"
          },
          "color": {
            "type": "string",
            "maxLength": 50,
            "title": "Color",
            "description": "Cor em formato OKLCH ou HEX"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "spent": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Spent",
            "default": "0.00"
          },
          "txs_count": {
            "type": "integer",
            "title": "Txs Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "name",
          "icon_name",
          "color",
          "id"
        ],
        "title": "CategoryResponse"
      },
      "CategoryUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "icon_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon Name"
          },
          "budget": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget"
          },
          "color": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "CategoryUpdate",
        "description": "Payload parcial do `PATCH /categories/{id}`.\n\n`extra=\"forbid\"` pelo mesmo motivo de `TransactionUpdate`: campo\ndesconhecido tem que falhar barulhento, não ser descartado."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InstallmentCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 100,
            "title": "Title"
          },
          "category_id": {
            "type": "integer",
            "title": "Category Id",
            "description": "ID da categoria cadastrada"
          },
          "total_amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Total Amount"
          },
          "installment_amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Installment Amount"
          },
          "current_installment": {
            "type": "integer",
            "title": "Current Installment"
          },
          "total_installments": {
            "type": "integer",
            "title": "Total Installments"
          },
          "end_date": {
            "type": "string",
            "title": "End Date"
          },
          "account_id": {
            "type": "integer",
            "title": "Account Id"
          }
        },
        "type": "object",
        "required": [
          "title",
          "category_id",
          "total_amount",
          "installment_amount",
          "current_installment",
          "total_installments",
          "end_date",
          "account_id"
        ],
        "title": "InstallmentCreate"
      },
...
      "InstallmentResponse": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 100,
            "title": "Title"
          },
          "category_id": {
            "type": "integer",
            "title": "Category Id",
            "description": "ID da categoria cadastrada"
          },
          "total_amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Amount"
          },
          "installment_amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Installment Amount"
          },
          "current_installment": {
            "type": "integer",
            "title": "Current Installment"
          },
          "total_installments": {
            "type": "integer",
            "title": "Total Installments"
          },
          "end_date": {
            "type": "string",
            "title": "End Date"
          },
          "account_id": {
            "type": "integer",
            "title": "Account Id"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "category": {
            "$ref": "#/components/schemas/CategoryRef"
          },
          "remaining_amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Remaining Amount",
            "description": "Quanto ainda falta pagar: parcela × parcelas restantes"
          }
        },
        "type": "object",
        "required": [
          "title",
          "category_id",
          "total_amount",
          "installment_amount",
          "current_installment",
          "total_installments",
          "end_date",
          "account_id",
          "id",
          "category",
          "remaining_amount"
        ],
        "title": "InstallmentResponse"
      },
...
      "InstallmentUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "category_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "total_amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Amount"
          },
          "installment_amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Installment Amount"
          },
          "current_installment": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Installment",
            "description": "Pode exceder total_installments — significa quitado"
          },
          "total_installments": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Installments"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "InstallmentUpdate",
        "description": "Payload parcial do `PATCH /installments/{id}`.\n\n`account_id` está fora do schema pela regra geral: IDs de relacionamento\ncentral são imutáveis via PATCH em toda a API. Com `extra=\"forbid\"` a\ntentativa vira 422 em vez de ser descartada em silêncio.\n\nA trava de `installment_amount`/`total_installments`/`total_amount` **não**\nestá aqui — ver `update_installment` no router. Ela depende de haver\ntransação vinculada e de o valor ter de fato mudado, dois fatos que só o\nbanco conhece."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}